Move is_default_cudf_dtype
